add product
Added user roles.
Different pages were redirected according to the logged in user.
If the user role is user, a redirect to the product add page has been made.
If the user role is member, the product details page has been redirected.
Description and picture added to product details section.