Add Category
Added category database.
Added category list page.
If the logged in person is "member", they are directed to the category list page.
Added "category id" column to products and associated with "id" in category table.
Whichever category was pressed in the category list, the products belonging to that category were listed on the product listing page.